Added support for using the VRAM as WRAM. Reworked the way in which the BGMAP segments are distributed between the param table and normal bgmaps. Renaming and cleaning up of libgccvb's old macros and globals.

+++ source/base/libgccvb/Affine.h
@@ -32,7 +32,7 @@
 // 											DECLARATIONS
 //---------------------------------------------------------------------------------------------------------
 
-fix19_13 Affine_applyAll(u16 param, fix19_13 paramTableRow, fix19_13 x, fix19_13 y, fix13_3 mx, fix13_3 my, fix19_13 halfWidth, fix19_13 halfHeight, const Scale* scale, const Rotation* rotation);
+fix19_13 Affine_applyAll(u32 param, fix19_13 paramTableRow, fix19_13 x, fix19_13 y, fix13_3 mx, fix13_3 my, fix19_13 halfWidth, fix19_13 halfHeight, const Scale* scale, const Rotation* rotation);
 
 
 #endif
